+++ backend/src/main/java/com/devsuperior/dsvendas/entities/Sale.java
@@ -11,8 +11,8 @@ public class Sale implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer visited;
-    private Integer deals;
+    private Long visited;
+    private Long deals;
     private Double amount;
     private LocalDate date;
 
@@ -23,7 +23,7 @@ public class Sale implements Serializable {
     public Sale() {
     }
 
-    public Sale(Long id, Integer visited, Integer deals, Double amount, LocalDate date, Seller seller) {
+    public Sale(Long id, Long visited, Long deals, Double amount, LocalDate date, Seller seller) {
         this.id = id;
         this.visited = visited;
         this.deals = deals;
@@ -40,19 +40,19 @@ public class Sale implements Serializable {
         this.id = id;
     }
 
-    public Integer getVisited() {
+    public Long getVisited() {
         return visited;
     }
 
-    public void setVisited(Integer visited) {
+    public void setVisited(Long visited) {
         this.visited = visited;
     }
 
-    public Integer getDeals() {
+    public Long getDeals() {
         return deals;
     }
 
-    public void setDeals(Integer deals) {
+    public void setDeals(Long deals) {
         this.deals = deals;
     }
